Introduce leaf certificate aliasing target
As of #37, tls-gen's basic profile includes hostname
into certificate and private key file names. I miss the
simplicify of having results/client_certificate.pem,
results/client_key.pem and so on.

So let's introduce a target that would compute the paths
and do the copying so I don't have to.